requirements.txt
oci
Flask
PyYAML
pandas
networkx
matplotlib